==> Step_3_run_models.ipynb <==
from statsmodels.iolib.summary2 import summary_col
import pandas as pd
import numpy as np
import statsmodels.api as sm

%store -r time_dummy
%store -r room_type_dummy

### read in cleaned data

df = pd.read_csv('airbnbdata_cleaned.csv', index_col = 0)
time_dummy = list(time_dummy.columns.values)
room_type_dummy = list(room_type_dummy.columns.values)
time_dummy = df[time_dummy]
room_type_dummy = df[room_type_dummy]

len(df.loc[df['review_scores_rating'] == 0, :])

len(df.loc[df['review_scores_rating'] == 1, :])

df.describe()['review_scores_rating']

### first try baseline model -- not a valid model

listing_attr = df[['property_type','accommodates','bathrooms','bedrooms', 'beds','amenities_counts','info_count']]
x_list = [listing_attr,room_type_dummy,time_dummy]
df_x1 = pd.concat(x_list,axis =1)
df_x_const = sm.add_constant(df_x1).apply(pd.to_numeric)

df_y = df['price in Euro']
modelbase = sm.OLS(df_y, df_x_const).fit() 
print(modelbase.summary())

### Model 1 with host attributies only 

listing_attr = df[['property_type','accommodates','bathrooms','bedrooms', 'beds','amenities_counts','info_count']]
x_list = [listing_attr,room_type_dummy,time_dummy]
df_x1 = pd.concat(x_list,axis =1)
df_x_const = sm.add_constant(df_x1).apply(pd.to_numeric)

df_y = df['price in Euro'].apply(lambda x: np.log(x))
model1 = sm.OLS(df_y, df_x_const).fit() 
print(model1.summary())

### Model-2 with listing attributes and distance to center  -- first we tried without normlization

dis_modle = df[['property_type','accommodates','bathrooms','bedrooms', 'beds','amenities_counts','info_count',\
                     'distance_Km']]
x_list = [dis_modle,room_type_dummy,time_dummy]
df_x2 = pd.concat(x_list,axis =1)
df_x_const = sm.add_constant(df_x2).apply(pd.to_numeric)

df_y = df['price in Euro'].apply(lambda x: np.log(x))
#df_y = df_listings['price in Euro']
model2 = sm.OLS(df_y, df_x_const).fit() 
print(model2.summary())

### Model-2 with listing attributes and distance to center - with normalization and got better result

df['distance_Km'] = df['distance_Km'].apply(lambda x: np.log(x))

dis_modle = df[['property_type','accommodates','bathrooms','bedrooms', 'beds','amenities_counts','info_count',\
                     'distance_Km']]
x_list = [dis_modle,room_type_dummy,time_dummy]
df_x2 = pd.concat(x_list,axis =1)
df_x_const = sm.add_constant(df_x2).apply(pd.to_numeric)

df_y = df['price in Euro'].apply(lambda x: np.log(x))
#df_y = df_listings['price in Euro']
model2 = sm.OLS(df_y, df_x_const).fit() 
print(model2.summary())

### Model-3 listing attributes+ distance to center+Policy Attributes

# instant_bookable true = 1 ## instant bookable ones has lower price cause those ones normamly want to attact users
#  strict = 1 strict has higher price 
# the beds one we can explain as special hosting or non special hosting
df3 = df[['minimum_nights','cancellation_policy','instant_bookable']]
df_x3 = pd.concat([df3,df_x2],axis =1)
df_x_const = sm.add_constant(df_x3).apply(pd.to_numeric)

df_y = df['price in Euro'].apply(lambda x: np.log(x))
#df_y = df_listings['price in Euro']
model3 = sm.OLS(df_y, df_x_const).fit() 
print(model3.summary())

### Model-4 listing attributes+ distance to center+Policy Attributes+reviews

df4 = df.dropna(subset = ['number_of_reviews','review_scores_rating'])
df4 = df4[['number_of_reviews','review_scores_rating']]
df_x4 = pd.concat([df4,df_x3],axis =1)
df_x_const = sm.add_constant(df_x4).apply(pd.to_numeric)
df_x_const = df_x_const.dropna(subset = ['number_of_reviews','review_scores_rating'])
df4index = list(df.index.values)

df_y = df['price in Euro'].apply(lambda x: np.log(x))
#df_y = df_listings['price in Euro']
model4 = sm.OLS(df_y, df_x_const).fit() 
print(model4.summary())

df_x4.columns.values

### Model-5 listing attributes+ distance to center+host attributes+Policy Attributes+reviews+hosts

df5 = df.dropna(subset = ['host_is_superhost','host_identity_verified'])
x_list = ['host_is_superhost','host_identity_verified']
df4list = list(df_x4.columns.values)
x_list.extend(df4list)
df_x5 = df5[x_list]
df_x_const = sm.add_constant(df_x5).apply(pd.to_numeric)

df_y = df5['price in Euro'].apply(lambda x: np.log(1+x))
#df_y = df_listings['price in Euro']
model5 = sm.OLS(df_y, df_x_const).fit() 
print(model5.summary())

info_dict={'R-squared' : lambda x: f"{x.rsquared:.6f}",
           'No. observations' : lambda x: f"{int(x.nobs):d}"}

results_table = summary_col(results=[model1,model2,model3,model4,model5],
                            float_format='%0.6f',
                            stars = True,
                            model_names=['Lisingt_attributes',
                                         'Add_distance',
                                         'Add_policy',
                                         'Add_review',
                                         'Add_host'],
                            info_dict=info_dict,
                            regressor_order = ['info_count',
                                             'amenities_counts',
                                             'beds',
                                             'bedrooms',
                                             'bathrooms',
                                             'accommodates',
                                             'property_type',
                                             'Private room',
                                             'Shared room',
                                             'distance_Km',
                                             'minimum_nights','cancellation_policy','instant_bookable',
                                             'number_of_reviews','review_scores_rating',
                                             'host_is_superhost','host_identity_verified'],
                           drop_omitted = True)
                                        

results_table.add_title('Airbnb-Price-Determinate-Regression')

print(results_table)
